# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Расчет времени расхождения языков методом глоттохронологии.",
        "title": "Language Annotation Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Вход в систему по логину и паролю. Возвращает JWT-токен и устанавливает сессионную куку.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Аутентификация пользователя",
                "parameters": [
                    {
                        "description": "Учетные данные пользователя",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "login": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "schema": {
                            "type": "object",
                            "properties": {
                                " jwt": {
                                    "type": "string"
                                },
                                " user": {
                                    "type": "object",
                                    "properties": {
                                        " is_linguist": {
                                            "type": "boolean"
                                        },
                                        " login": {
                                            "type": "string"
                                        },
                                        "id": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Неверный логин или пароль",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Удаляет сессию пользователя и очищает cookie.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Выход из системы",
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/users/register": {
            "post": {
                "description": "Создает нового пользователя с логином и паролем.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "user_credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "login": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешная регистрация",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверные данные или пользователь уже существует",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Введите токен JWT (например: Bearer \u003ctoken\u003e)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "Введите куку сессии (например: session_id=\u003cvalue\u003e)",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}